+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 - Scopes can build an hierrarchy
+- Hand over components and previous value to produce method
 
 ## [1.0.0] - 2024-05-04